Deconstruct splash background box composable
This is necessary to change the structure a bit: the buttons composable
must be a sibling to the clipped background to utilize subcomposition.
