+++ WordPress/src/jetpack/java/org/wordpress/android/ui/accounts/login/LoginPrologueRevampedFragment.kt
@@ -7,24 +7,45 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.view.Window
 import android.view.WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.BlurredEdgeTreatment
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.draw.blur
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.geometry.Rect
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.Outline.Rectangle
+import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.Density
+import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import androidx.fragment.app.Fragment
+import org.wordpress.android.R.drawable
+import org.wordpress.android.R.string
 import org.wordpress.android.ui.accounts.login.components.ButtonsColumn
 import org.wordpress.android.ui.accounts.login.components.JetpackLogo
 import org.wordpress.android.ui.accounts.login.components.PrimaryButton
 import org.wordpress.android.ui.accounts.login.components.SecondaryButton
-import org.wordpress.android.ui.accounts.login.components.SplashBackgroundBox
+import org.wordpress.android.ui.accounts.login.components.SplashBox
 import org.wordpress.android.ui.compose.theme.AppTheme
 import org.wordpress.android.util.extensions.showFullScreen
 
@@ -81,16 +102,48 @@ private fun LoginScreenRevamped(
 ) {
     val blurredAreaHeight = remember { mutableStateOf(0) }
 
-    SplashBackgroundBox(blurredAreaHeight) {
-        JetpackLogo(
-                modifier = Modifier
-                        .padding(top = 60.dp)
-                        .size(60.dp)
+    Box {
+        val blurClipShape = remember {
+            object : Shape {
+                override fun createOutline(size: Size, layoutDirection: LayoutDirection, density: Density): Rectangle {
+                    return Rectangle(
+                            Rect(
+                                    bottom = size.height,
+                                    left = 0f,
+                                    right = size.width,
+                                    top = size.height - blurredAreaHeight.value,
+                            )
+                    )
+                }
+            }
+        }
+
+        SplashBox()
+        SplashBox(
+                modifier = Modifier.clip(blurClipShape),
+                textModifier = Modifier
+                        .blur(15.dp, BlurredEdgeTreatment.Unbounded)
+                        .alpha(0.5f) // Fallback for Android versions older than 12 where blur is not supported
         )
-        Spacer(Modifier.weight(1.0f))
-        ButtonsColumn(Modifier.onSizeChanged { blurredAreaHeight.value = it.height }) {
-            PrimaryButton(onClick = onWpComLoginClicked)
-            SecondaryButton(onClick = onSiteAddressLoginClicked)
+        Image(
+                painter = painterResource(drawable.bg_jetpack_login_splash_top_gradient),
+                contentDescription = stringResource(string.login_prologue_revamped_content_description_top_bg),
+                contentScale = ContentScale.FillBounds,
+                modifier = Modifier
+                        .fillMaxWidth()
+                        .height(height = 292.dp)
+        )
+        Column(horizontalAlignment = Alignment.CenterHorizontally) {
+            JetpackLogo(
+                    modifier = Modifier
+                            .padding(top = 60.dp)
+                            .size(60.dp)
+            )
+            Spacer(Modifier.weight(1.0f))
+            ButtonsColumn(Modifier.onSizeChanged { blurredAreaHeight.value = it.height }) {
+                PrimaryButton(onClick = onWpComLoginClicked)
+                SecondaryButton(onClick = onSiteAddressLoginClicked)
+            }
         }
     }
 }
